+++ bin-proto/src/impls/container.rs
@@ -1,8 +1,11 @@
+#[allow(unused)]
 macro_rules! impl_container_write {
     (
+        $(#[$attr:meta])?
         $ty:ident<$($a:lifetime,)? T $(: $tbound0:ident $(+ ?$tbound1:ident + $tbound2:lifetime)?)?>
         $(=> $f:ident)?
     ) => {
+        $(#[$attr])?
         impl<$($a,)? Ctx, Tag, T> $crate::BitEncode<Ctx, Tag> for $ty<$($a,)? T>
         where
             T: $crate::BitEncode<Ctx, Tag> $(+ $tbound0 $(+ ?$tbound1 + $tbound2)?)?,
@@ -30,8 +33,10 @@ macro_rules! impl_container_write {
     };
 }
 
+#[allow(unused)]
 macro_rules! impl_container_read {
-    ($ty:ident<T>) => {
+    ($(#[$attr:meta])? $ty:ident<T>) => {
+        $(#[$attr])?
         impl<Ctx, Tag, T> $crate::BitDecode<Ctx, Tag> for $ty<T>
         where
             T: $crate::BitDecode<Ctx, Tag>,
@@ -47,37 +52,41 @@ macro_rules! impl_container_read {
     };
 }
 
+#[cfg(feature = "alloc")]
 mod box_ {
     use alloc::boxed::Box;
 
-    impl_container_write!(Box<T>);
-    impl_container_read!(Box<T>);
+    impl_container_write!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Box<T>);
+    impl_container_read!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Box<T>);
     test_codec!(Box<u8>; Box::new(1) => [0x01]);
     test_roundtrip!(Box<u8>);
 }
 
+#[cfg(feature = "alloc")]
 mod rc {
     use alloc::rc::Rc;
 
-    impl_container_write!(Rc<T>);
-    impl_container_read!(Rc<T>);
+    impl_container_write!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Rc<T>);
+    impl_container_read!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Rc<T>);
     test_codec!(Rc<u8>; Rc::new(1) => [0x01]);
     test_roundtrip!(Rc<u8>);
 }
 
+#[cfg(feature = "alloc")]
 mod arc {
     use alloc::sync::Arc;
 
-    impl_container_write!(Arc<T>);
-    impl_container_read!(Arc<T>);
+    impl_container_write!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Arc<T>);
+    impl_container_read!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Arc<T>);
     test_codec!(Arc<u8>; Arc::new(1) => [0x01]);
     test_roundtrip!(Arc<u8>);
 }
 
+#[cfg(feature = "alloc")]
 mod cow {
     use alloc::borrow::{Cow, ToOwned};
 
-    impl_container_write!(Cow<'a, T: ToOwned + ?Sized + 'a>);
+    impl_container_write!(#[cfg_attr(docsrs, doc(cfg(feature = "alloc")))] Cow<'a, T: ToOwned + ?Sized + 'a>);
     test_encode!(Cow<u8>; Cow::Owned(1) => [0x01]);
 }