+++ bin-proto/src/impls/str_.rs
@@ -1,6 +1,5 @@
 use crate::{util, BitEncode, Result};
 
-use alloc::vec::Vec;
 use bitstream_io::{BitWrite, Endianness};
 
 impl<Ctx> BitEncode<Ctx> for str {
@@ -9,8 +8,7 @@ impl<Ctx> BitEncode<Ctx> for str {
         W: BitWrite,
         E: Endianness,
     {
-        let bytes: Vec<_> = self.bytes().collect();
-        util::encode_items::<_, E, _, _>(&bytes, write, ctx)
+        util::encode_items::<_, E, _, _>(self.as_bytes(), write, ctx)
     }
 }