+++ bin-proto/src/util.rs
@@ -43,6 +43,7 @@ where
     Ok(())
 }
 
+#[allow(unused)]
 pub fn decode_items_to_eof<'a, R, E, Ctx, T>(
     read: &'a mut R,
     ctx: &'a mut Ctx,